Replace unused endpoint, add project README, update test.http
Removed the unimportant create_file endpoint from FakeCDN/upload.py to make the API more clear. Added a README.md file to provide documentation on how to run the application and which endpoints can be used for testing. Modified the test.http file to add placeholders for the image path to make the testing process more user-friendly.