Update README.md for clarification
Changes were made in README.md to provide more accurate information about the repository. "This repo is a reference to the NFT project." was changed to "This repo is a basically a playground for me to learn FastAPI and would be used in the [NFT-Backend](https://github.com/Dragonfly911117/NFT-Backend) project", providing specific clarity regarding the role of this repository within the context of the larger NFT-Backend project.