Refine project introduction in README.md
The project description in README.md file has been refined to clearly elucidate the project's purpose. The narrative was a bit ambiguous as to what role this repository plays, which has now been modified to "This repo is basically a playground for me to learn FastAPI and would be used in the [NFT-Backend](https://github.com/Dragonfly911117/NFT-Backend) project". This change is intended to provide clearer context for any future project contributors and viewers.